airflow/requirements.txt
requests
xlrd
openpyxl
pyarrow
apache-airflow-providers-google
pyarrow

